+++ tests/domain/business/candidate/test_sortered_candidate.py
@@ -2,11 +2,9 @@ import unittest
 import datetime
 
 from app.domain.business.candidate.sortered_candidate import SorteredCandidate
-from app.domain.business.party_business import PartyBusiness
-from app.domain.repository.candidate_repository import CandidateRepository
 from tests.assert_test import AssertTest
-from tests.faker import getCandidates, getParties
-from unittest.mock import patch
+from tests.faker.faker_party import getParties_by_id
+from tests.faker.faker_candidate import getCandidates_by_id
 
 
 class SorteredCandidateTest(unittest.TestCase) : 
@@ -16,8 +14,8 @@ class SorteredCandidateTest(unittest.TestCase) :
         
         
     def test_get_top_candidates_results_small_samples(self):
-        candidates = getCandidates([2, 8, 14, 26, 28])
-        parties = getParties([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
+        candidates = getCandidates_by_id([2, 8, 14, 26, 28])
+        parties = getParties_by_id([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
         
         business = SorteredCandidate(candidates, parties)
         
@@ -57,9 +55,9 @@ class SorteredCandidateTest(unittest.TestCase) :
           
         
     def test_get_top_candidates_results(self):
-        candidates = getCandidates([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 
+        candidates = getCandidates_by_id([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 
                                                                 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35])
-        parties= getParties([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
+        parties= getParties_by_id([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])
         
         business = SorteredCandidate(candidates, parties)
