+++ tests/domain/business/test_department_business.py
@@ -5,7 +5,9 @@ from app.domain.repository.department_repository import DepartmentRepository
 from app.domain.repository.district_repository import DistrictRepository
 from app.domain.repository.result_repository import ResultRepository
 from tests.assert_test import AssertTest
-from tests.faker import getDepartments, getDistricts, getResults
+from tests.faker.faker_result import getResults
+from tests.faker.faker_district import getDistricts_by_id
+from tests.faker.faker_department import getDepartments_by_ids
 from unittest.mock import patch
 
 class DepartmentBusinessTest(unittest.TestCase):
@@ -16,7 +18,7 @@ class DepartmentBusinessTest(unittest.TestCase):
     
     @patch.object(DepartmentRepository, "get_departments")
     def test_get_departments(self, mock_repo):
-        mock_repo.get_departments.return_value = getDepartments([1, 2, 3, 33])
+        mock_repo.get_departments.return_value = getDepartments_by_ids([1, 2, 3, 33])
         
         business = DepartmentBusiness(mock_repo, None, None)
         
@@ -39,7 +41,7 @@ class DepartmentBusinessTest(unittest.TestCase):
         
     @patch.object(DepartmentRepository, "get_departments")
     def test_get_departments_by_name(self, mock_repo):
-        mock_repo.get_departments.return_value = getDepartments([1, 2, 3, 33])
+        mock_repo.get_departments.return_value = getDepartments_by_ids([1, 2, 3, 33])
         
         business = DepartmentBusiness(mock_repo, None, None)
         
@@ -51,7 +53,7 @@ class DepartmentBusinessTest(unittest.TestCase):
 
     @patch.object(DepartmentRepository, "get_departments")
     def test_get_department_name_from_department_id(self, mock_dep_repo): 
-        mock_dep_repo.get_departments.return_value = getDepartments([1, 2, 3, 33])
+        mock_dep_repo.get_departments.return_value = getDepartments_by_ids([1, 2, 3, 33])
         
         business = DepartmentBusiness(mock_dep_repo, None, None)
         
@@ -64,8 +66,8 @@ class DepartmentBusinessTest(unittest.TestCase):
     @patch.object(DistrictRepository, "get_districts")
     @patch.object(ResultRepository, "get_results")
     def test_get_departments_by_voting_rate_ascending(self, mock_department_repository, mock_district_repository, mock_result_business):
-        mock_department_repository.get_departments.return_value = getDepartments([1, 11, 33, 34, 69])
-        mock_district_repository.get_districts.return_value = getDistricts([10, 11, 12, 110, 111, 330, 331, 340, 341, 342, 343, 690, 691, 692, 693])
+        mock_department_repository.get_departments.return_value = getDepartments_by_ids([1, 11, 33, 34, 69])
+        mock_district_repository.get_districts.return_value = getDistricts_by_id([10, 11, 12, 110, 111, 330, 331, 340, 341, 342, 343, 690, 691, 692, 693])
         mock_result_business.get_results.return_value =  getResults([100, 105, 110, 115, 120, 125, 1100, 1105, 1110, 1115, 3300, 3305, 3310, 3315, 
                                                                      3400, 3405, 3410, 3415, 3420, 3425, 3430, 3435, 6900, 6905, 6910, 6915, 6920, 6925, 6930, 6935])    
                
@@ -97,8 +99,8 @@ class DepartmentBusinessTest(unittest.TestCase):
     @patch.object(DistrictRepository, "get_districts")
     @patch.object(ResultRepository, "get_results")    
     def test_get_departments_by_voting_rate_with_creuse_have_only_district_and_one_result(self, mock_department_repository, mock_district_repository, mock_result_business):
-        mock_department_repository.get_departments.return_value = getDepartments([1, 11])
-        mock_district_repository.get_districts.return_value = getDistricts([10, 11, 12, 110, 111])
+        mock_department_repository.get_departments.return_value = getDepartments_by_ids([1, 11])
+        mock_district_repository.get_districts.return_value = getDistricts_by_id([10, 11, 12, 110, 111])
         mock_result_business.get_results.return_value =  getResults([100, 105, 110, 115, 120, 125, 1105, 1115])
         
         department_business = DepartmentBusiness(mock_department_repository, mock_district_repository, mock_result_business)
@@ -122,8 +124,8 @@ class DepartmentBusinessTest(unittest.TestCase):
     @patch.object(DistrictRepository, "get_districts")
     @patch.object(ResultRepository, "get_results")
     def test_get_departments_by_voting_rate_descending(self, mock_department_repository, mock_district_repository, mock_result_business):
-        mock_department_repository.get_departments.return_value = getDepartments([1, 11, 33, 34, 69])
-        mock_district_repository.get_districts.return_value = getDistricts([10, 11, 12, 110, 111, 330, 331, 340, 341, 342, 343, 690, 691, 692, 693])
+        mock_department_repository.get_departments.return_value = getDepartments_by_ids([1, 11, 33, 34, 69])
+        mock_district_repository.get_districts.return_value = getDistricts_by_id([10, 11, 12, 110, 111, 330, 331, 340, 341, 342, 343, 690, 691, 692, 693])
         mock_result_business.get_results.return_value =  getResults([100, 105, 110, 115, 120, 125, 1100, 1105, 1110, 1115, 3300, 3305, 3310, 3315, 
                                                                      3400, 3405, 3410, 3415, 3420, 3425, 3430, 3435, 6900, 6905, 6910, 6915, 6920, 6925, 6930, 6935])
